src/main/java/com/felix/common/collection/Collections.java
package com.felix.common.collection;

import java.util.Collection;
import java.util.Objects;
import java.util.Set;
import java.util.concurrent.ConcurrentHashMap;
import java.util.function.Function;
import java.util.function.Predicate;

public class Collections {

    public static boolean isNullOrEmpty(Collection<?> c) {
        return c == null || c.isEmpty();
    }

    public static <T> boolean anyMatch(Collection<T> c, Predicate<T> predicate) {
        Objects.requireNonNull(c);

        return c.stream().anyMatch(predicate);
    }

    public static <T> Predicate<T> distinctBy(Function<? super T, ?> keyExtractor) {
        Set<Object> seen = ConcurrentHashMap.newKeySet();
        return t -> seen.add(keyExtractor.apply(t));
    }
}
